requirements.txt
alpaca-py>=0.21.0
requests>=2.31.0
python-dotenv>=1.0.0
pandas>=2.0.0
numpy>=1.24.0
scipy>=1.10.0
